upgrade codebase to rails 4.x
Move to Rails 4.1.4 and the latest available Gems for the whole
environment. Also move all secrets into the new secrets file and
make all other files read their secrets from there instead.

In addition, expand the repository to include all but private
files which should not be on the repo.