create a new search engine
Start work on a new search engine. The search engine will take queries, save them,
pass them onto backends (Last.fm, SoundCloud, YouTube, etc.), parse the results
and present it back to the user.

If a search is made via the API (json or xml) the results are fetched directly and
simply returned.

If, however, the search is done interactively (a normal, browsing user) the user is
first presented with a skeleton result page and the query is sent via AJAX and
injected onto the page.