search: add fullname field for better search ranking
One of the parameters the search engine uses for ranking is similarity
between the searched query and the name of the hit. For songs the displayed
name is just the title (if the parsing did a good job).

So we add a new field to each hit called :fullname which contains something
like "Artist - Song Title" and run the similarity calculator on this field
instead of the displayed string (which would be "Song Title" in our example).

This will increase the similarity value for songs made by the searched artist
which should produce better results.