website: create source object for linking songs to external websites
One song can appear on both YouTube and SoundCloud, and there may
even be duplicates within those websites. So I decided to create a
Source object which contains information about the source website
(id, url, etc.) and this object can then be assigned to a song
(or any other resource which may have links to external websites).

Since the relation is one-to-many this means we can eliminate duplicates
and collect all sources for a song in a single place.