website: add image object
Since resouces such as songs, artists and albums can have more than one image attached to them,
often the same art but in different sizes, it makes sense to create an image object and attach
it to various resources using a one-to-many relationship. This image can point to an external
url or a relative path, and supply a size.