website: implement searching inside database
Search results from backends are now saved in the database. If the
same search query is executed within a week the search controller
will instead search the database instead of the backends. This should
speed up searching and minimize our requests to the various APIs.

The database searching is powered by Solr via the Sunspot gem.