website: restructure hash with search results
There is really no need to put search results into [:object], instead
the :hits array is plainly an array of objects.

Similarly the :exact value is a single object, and the controller is
given the power to choose which object should be determined an exact
match (instead of the view).