website: start work on rewriting all html/js/css
The actual web interface needs to be rewritten. First and foremost
it needs to be responsive. It also needs to be more structured and
heavily refactored.

This first step removes all the old CSS and replaces it with new
stylesheets. The HTML of the layout files has also been replaced
to use more HTML5 and make the markup a lot cleaner and lighter.

This will of course break almost everything. Right now, no actual
content is yielded in the layout. It only contains a header, sidebar,
footer and some static content.

The next step is to one-by-one go through each page and fit it into
the new site.