website: start move to responsive design
Responsive design is all the rage. Mostly because it allows you to
have a single website working across multiple screen sizes. Simply
speaking it is an initial investment which pays off later.

This commit starts this conversion. The following pages are
converted:
* First page
* Get/download
* News
* Tour
* Legal
* Money

In this move we also move fully to CoffeeScript and restructure all
Javascript and CSS into a more logical structure.