website: remove old, lingering flash messages
flash messages (alerts and notices) are shown on the login form even
after several page views after they have been created. For example if
you log out, then visit pages which do not display the flash messages,
and then visit the login page, you are greeted by a message telling you
that you have been logged out.

It seems that the flash message that Devise sets at log out
is only cleared when it is displayed (or discarded manually).

The fix for this is to discard all flash messages at the footer of the page.
Then each page view will clear the flash at the end of the rendering.
Redirects will keep the flash, so it is only discarded when a page is
rendered (actually, when the shared/_footer partial is rendered).