website: improve detection of old browser
Use Webkit version instead of Safari version to determine if the
browser is old or not. This improves detection since some browser (such
as Chrome on iPad) will report as Webkit but will not contain a Safari
version.

The current code considers Safari 7 and above as modern enough, which
corresponds to Webkit version 537. However, on Windows the latest
version is 534 so a lower level of tolerance is used when the platform
is detected as Windows.