website: make user menu in navigation responsive
Change the user menu in the navigation bar when the screen is small.
Instead of displaying the user’s photo and name, just show a link to
dashboard.

Remove the link tag on the menu button so that it can be pressed on
touch devices to toggle the dropdown menu.