website: use strong parameters in link controller
New in Rails is the usage of “strong parameters”. Update the link
controller to do this.

Also change redirection to account tab on settings page, instead of
just the settings page.

Lastly, fix a bug that tries to render a non-existing page when
supposed to render forbidden. Fix it to just return headers and a
forbidden status.